refactor(scheduling): Refactors
- Refactored scheduling components to individual files
