--- src/modules/PushAnalytics/components/ScheduleModal/components/ScheduleLabel.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import {useWatch} from "react-hook-form";
-
-
-export function ScheduleLabel() {
-    const data = useWatch();
-
-    return (
-        <div>
-
-        </div>
-    )
-}
